Updated Testing Framework and optimized filter process
Future steps are now pretty much only optimizing parameters. We could also use these methods as similarity edges for a clustering method of which there are many (Random Walk Paper). I also need to more thouroughly understand the analysis of each filter so that I can decide which and how many layers of AND/OR blocks we should use. Chaining filters together too can give each one a role which will be useful.