Added ROC curve generation
Currently getting some strange results. I'm not exactly sure what the cause is. I'm hoping it's just a bug in my code but it could also be something strange with analyzing all regions as opposed to the subset. I will compare the method in the notebook to the command line to make sure we are reproducing the same results.